+++ product/dashboard/src/webapp/src/app/pages/collection-status/collection-status.component.state.ts
@@ -2,8 +2,6 @@ import {AgePipe} from '../../pipes/age.pipe';
 import {DashboardApiService} from '../../services/dashboard-api.service';
 import {StatusData} from '../../model/status/StatusData';
 import {Subscription, timer} from 'rxjs';
-import {Agent} from '../../model/status/Agent';
-import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
 export class CollectionStatusComponentState {
     static KEY = 'collection-status';
@@ -13,8 +11,6 @@ export class CollectionStatusComponentState {
     autoRefresh = true;
     showTerminatedAgents = false;
     refreshIntervalSeconds = 60;
-    selectAllTerminatedAgents = false;
-    selectTerminatedAgentsOlderThanDays = 30;
 
     applicationFilter = '';
     environmentFilter = '';
@@ -22,7 +18,7 @@ export class CollectionStatusComponentState {
 
     private timerSubscription: Subscription;
 
-    constructor(private agePipe: AgePipe, private api: DashboardApiService, private modalService: NgbModal) {
+    constructor(private agePipe: AgePipe, private api: DashboardApiService) {
     }
 
     init() {
@@ -73,11 +69,9 @@ export class CollectionStatusComponentState {
             .subscribe(data => {
                 this.data = data;
                 this.errorMessage = undefined;
-                this.selectAllTerminatedAgents = false;
             }, error => {
                 this.data = undefined;
                 this.errorMessage = error.statusText ? error.statusText : error;
-                this.selectAllTerminatedAgents = false;
             });
     }
 
@@ -131,63 +125,6 @@ export class CollectionStatusComponentState {
         return null;
     }
 
-    numSelectedTerminatedAgents() {
-        if (this.data.agents) {
-            return this.data.agents.filter(a => a.selected).length;
-        }
-        return null;
-    }
-
-    numSelectedVisibleTerminatedAgents() {
-        if (this.getVisibleAgents()) {
-            return this.getVisibleAgents().filter(a => a.selected).length;
-        }
-        return null;
-    }
-
-    selectOrUnselectAllVisibleTerminatedAgents() {
-        if (this.getVisibleAgents()) {
-            if (this.autoRefresh) {
-                this.toggleAutoRefresh();
-            }
-            this.getVisibleAgents()
-                .forEach(
-                    a => a.selected = this.selectAllTerminatedAgents
-                        && !a.agentAlive
-                        && (a.deletionState === null || a.deletionState === undefined));
-        }
-    }
-
-    getTerminatedBefore(): Date {
-        let d = new Date();
-        d.setDate(d.getDate() - this.selectTerminatedAgentsOlderThanDays);
-        return d;
-    }
-
-    selectOldTerminatedAgents() {
-        if (this.data.agents) {
-            if (this.autoRefresh) {
-                this.toggleAutoRefresh();
-            }
-            this.data.agents.filter(a => !a.agentAlive && a.publishedAtMillis < this.getTerminatedBefore().getTime())
-                .forEach(a => a.selected = true);
-        }
-    }
-
-    openDeleteModal(content: any) {
-        this.modalService.open(content).result.then(() => {
-            this.deleteSelectedAgents();
-        }, () => {
-
-        });
-    }
-
-    deleteSelectedAgents() {
-        if (this.data.agents) {
-            this.data.agents.filter(a => a.selected && !a.agentAlive).forEach(a => this.deleteAgent(a));
-        }
-    }
-
     private startAutoRefresh() {
         this.timerSubscription = timer(0, this.refreshIntervalSeconds * 1000).subscribe((tick: number) => {
             console.log('[ck dashboard] Doing auto-refresh #%o', tick);
@@ -199,11 +136,4 @@ export class CollectionStatusComponentState {
         this.timerSubscription.unsubscribe();
     }
 
-    private deleteAgent(agent: Agent) {
-        agent.deletionState = 1;
-        this.api.deleteAgent(agent.agentId, agent.jvmId).subscribe(() => {
-            agent.deletionState = 2;
-            agent.selected = false;
-        });
-    }
 }