+++ product/dashboard/src/webapp/src/app/services/dashboard-api.service.ts
@@ -11,7 +11,6 @@ import {StatusData} from '../model/status/StatusData';
 @Injectable()
 export class DashboardApiService {
 
-    readonly DELETE_AGENT_URL = '/dashboard/api/v1/agent';
     readonly METHOD_BY_ID_URL = '/dashboard/api/v1/method/detail/';
     readonly METHODS_FORM_DATA_URL = '/dashboard/api/v1/methodsFormData';
     readonly METHODS_URL = '/dashboard/api/v2/methods';
@@ -44,9 +43,5 @@ export class DashboardApiService {
 
     getServerSettings() {
         return this.http.get<ServerSettings>(this.SERVER_SETTINGS_URL);
-    }
-
-    deleteAgent(agentId: number, jvmId: number) {
-        return this.http.delete(`${this.DELETE_AGENT_URL}/${agentId}/${jvmId}`);
     }
 }